requirements.txt
openai
langgraph
langchain_openai
langchain_core
google-api-python-client
google-auth-httplib2
google-auth-oauthlib==0.8.0
google-auth
